Now it is working.I did a general refactoring by removing the Score class, the live field of the Stone class, I also created a FreedomPointsCounterTests class which counts the points of the players.
