+++ src/main/java/Game.java
@@ -1,4 +1,3 @@
-import exceptions.InvalidPositionException;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -19,16 +18,10 @@ public class Game {
     private final LinkedList<Move> allPlayersMoves = new LinkedList<>();
     private final TextInput userInput = new TextInput();
 
-    private final Score whiteScore;
-
-    private final Score blackScore;
-
     public Game(@NotNull Board board, @NotNull Player whitePlayer, @NotNull Player blackPlayer) {
         this.whitePlayer = whitePlayer;
         this.blackPlayer = blackPlayer;
         this.board = board;
-        this.whiteScore = new Score(this.whitePlayer);
-        this.blackScore = new Score(this.blackPlayer);
     }
 
     public void start() {
@@ -38,9 +31,12 @@ public class Game {
         while (gameStatus != GameStatus.GAME_OVER) {
             turn();
         }
-        System.out.println("The winner is: " + getTheWinner());
-        System.out.println("White score: " + whiteScore.getNumberOfFreedomLines());
-        System.out.println("Black score" + blackScore.getNumberOfFreedomLines());
+        Player winner = getTheWinner();
+        if (winner != null) {
+            System.out.println("The winner is: " + getTheWinner());
+        } else {
+            System.out.println("Tie!");
+        }
         end();
     }
 
@@ -172,88 +168,17 @@ public class Game {
         }
     }
 
-    @NotNull
+    @Nullable
     public Player getTheWinner() {
-        for (Map.Entry<Position, Cell> entry : board) {
-            Position currentPosition = entry.getKey();
-            checkFreedomLineFrom(currentPosition);
-        }
-        if (blackScore.getNumberOfFreedomLines() > whiteScore.getNumberOfFreedomLines()) {
+        FreedomPointsCounter freedomPointsCounter = new FreedomPointsCounter(board);
+        freedomPointsCounter.count();
+        if (freedomPointsCounter.getWhitePlayerScore() > freedomPointsCounter.getBlackPlayerScore()) {
+            return whitePlayer;
+        } else if (freedomPointsCounter.getBlackPlayerScore() > freedomPointsCounter.getWhitePlayerScore()) {
             return blackPlayer;
         }
-        return whitePlayer;
+        return null;
     }
-
-
-    private void checkFreedomLineFrom(@NotNull Position position) {
-        if (!board.isCellOccupied(position)) return;
-        Stone.Color playerColor = board.getStone(position).getColor();
-
-        if (countStonesOfTheSameColor(position,1,"HORIZONTAL") == 4) {
-            incrementScoreOf(playerColor);
-        }
-        if (countStonesOfTheSameColor(position,1,"VERTICAL") == 4) {
-            incrementScoreOf(playerColor);
-        }
-        if (countStonesOfTheSameColor(position,1,"DIAGONAL-LEFT") == 4) {
-            incrementScoreOf(playerColor);
-        }
-        if (countStonesOfTheSameColor(position,1,"DIAGONAL-RIGHT") == 4) {
-            incrementScoreOf(playerColor);
-        }
-
-    }
-
-    private void incrementScoreOf(Stone.Color playerColor) {
-        if (playerColor == Stone.Color.BLACK) {
-            blackScore.incrementNumberOfFreedomLines();
-        } else {
-            whiteScore.incrementNumberOfFreedomLines();
-        }
-    }
-
-    private int countStonesOfTheSameColor(@NotNull Position currentPosition, int currentStoneCount, String direction) {
-        Position nextPosition;
-        if(getThePreviousStoneColor(currentPosition,direction) == board.getStone(currentPosition).getColor()) {
-            return currentStoneCount;
-        }
-        try {
-            switch (direction) {
-                case "HORIZONTAL":
-                    nextPosition = new Position(currentPosition.getRow(), currentPosition.getColumn() + 1);
-                    break;
-                case "VERTICAL":
-                    nextPosition = new Position(currentPosition.getRow() + 1, currentPosition.getColumn());
-                    break;
-                case "DIAGONAL-LEFT":
-                    nextPosition = new Position(currentPosition.getRow() + 1, currentPosition.getColumn() - 1);
-                    break;
-                default:
-                    nextPosition = new Position(currentPosition.getRow() + 1, currentPosition.getColumn() + 1);
-            }
-        } catch (InvalidPositionException exception) {
-            return currentStoneCount;
-        }
-        if (board.getStone(currentPosition).getColor() == board.getStone(nextPosition).getColor()) {
-            return countStonesOfTheSameColor(nextPosition, currentStoneCount + 1, direction);
-        } else {
-            return currentStoneCount;
-        }
-    }
-
-    private Stone.Color getThePreviousStoneColor(Position currentPosition, String direction) throws InvalidPositionException {
-        switch (direction) {
-            case "HORIZONTAL":
-                return board.getStone(new Position(currentPosition.getRow(), currentPosition.getColumn() - 1)).getColor();
-            case "VERTICAL":
-                return board.getStone(new Position(currentPosition.getRow() - 1, currentPosition.getColumn())).getColor();
-            case "DIAGONAL-LEFT":
-                return board.getStone(new Position(currentPosition.getRow() - 1, currentPosition.getColumn() + 1)).getColor();
-            default:
-                return board.getStone(new Position(currentPosition.getRow() - 1, currentPosition.getColumn() - 1)).getColor();
-        }
-    }
-
 
 
 }
